Seperate active tasks and completed tasks
Big change, I added a seperate file for completed tasks, so that the
primary application is not bogged down by a mass amount of completed
tasks. Completed tasks have their own logic and functions. This is a
very iterative change, and will require a good amount of work to
implement.